refactor(proxy,http): move proxy_server from mockforge-http to mockforge-proxy (#555 phase 1)
ADR 0001 marked mockforge-http/src/proxy_server.rs as a "Clean move"
candidate to mockforge-proxy: its only foreign imports were already in
mockforge-proxy (body_transform, config::ProxyConfig), and the only
workspace caller was tests/proxy_verification_tests.rs.

## What changed

- `git mv` crates/mockforge-http/src/proxy_server.rs →
  crates/mockforge-proxy/src/server.rs.
- Rewrote the file's `mockforge_proxy::*` imports to `crate::*`.
- Added `pub mod server;` to mockforge-proxy/src/lib.rs.
- mockforge-http/src/proxy_server.rs is now a 1-line re-export shim so
  existing callers (proxy_verification_tests + any out-of-tree
  consumers) keep resolving.
- mockforge-proxy/Cargo.toml: added tokio (sync/macros/rt/time) and
  http = "1.3" (matching mockforge-http's pin) — the moved code reaches
  http::{Uri, Request} directly, not via axum's re-export.

## Verification

- cargo check -p mockforge-proxy -p mockforge-http — clean
- cargo test -p mockforge-proxy --lib — 17 pass (3 ProxyServer tests inherited)
- cargo test --test proxy_verification_tests — 2 pass (exercises the shim)
- cargo clippy -p mockforge-proxy -p mockforge-http --all-targets -- -D warnings — clean
- cargo fmt --all --check — clean
